Help with Code after Project Sumbmission
I need help with the two functions in the body of the code which my
reviewer suggested I add to remove redundant code. I can’t call them
properly. I fixed constraints. The view is shifting down and not up for
some reason when a text field becomes first responder.